# Fix interoperability with native async generators