- Revigorate support for using the native AsyncGenWrappedValue type as our wrapper type on CPython 3.6+. Instead of trying to manually construct the object using ctypes, trick the interpreter into doing the wrapping/unwrapping for us.
- Add support for PEP 525 async generator hooks (like `sys.set_asyncgen_hooks()`); on 3.6+, our hooks are the same as the system ones, and regardless of version they can be accessed via `async_generator.set_asyncgen_hooks()` and so on
- Add the ability for `@async_generator` to produce a native async generator. Automatically detect whether this is possible (we need to be on CPython 3.6+ and the async function we were given needs to only ever return None). Since there are a few minor semantic differences, native mode currently must be requested using `@async_generator_native`, and plain `@async_generator` warns on code whose behavior would differ with _native. In a future release we can make _native the default, and require people who really need ag_running to work correctly (for example) to use `@async_generator_legacy`.