@oremanj 2018-03-02
I didn't have a specific reason at first - I came across the commented-out code and thought, "this sounds fun to work out!", so I did.

Thinking about it some more... I think the only real utility is if we could use this to make @async_generator automatically produce a native async generator on 3.6+, which would have both interoperability and performance improvements. Interoperability is of limited use given that native async generators can't use "yield from", and we may or may not care about the performance - I know your feelings about microbenchmarks. :-) Since some numbers are better than none, though, I extended the microbenchmark from PEP 525 to compare the analogous @async_generator and two ways I thought of for making a looks-like-an-@async_generator-but-is-actually-native:
```
N = 10 ** 7

from async_generator import async_generator, yield_

# iterate over an async iterator and return the sum of the values provided
async def test(ait):
    sum = 0
    async for i in ait:
        sum += i
    return sum

# run a secretly-sync coroutine
def run(coro):
    try:
        coro.send(None)
    except StopIteration as ex:
        return ex.value
    else:
        raise RuntimeError("wasn't sync")  

# native async generator
async def agen():
    for i in range(N):
        yield i

# current @async_generator without ctypes hackery
@async_generator
async def agen2():
    for i in range(N):
        await yield_(i)

# hypothetical way of making a native asyncgen out of an async function that
# yields values by await yield_native(val)
import types, ctypes, functools
dll = ctypes.CDLL(None)
dll._PyAsyncGenValueWrapperNew.restype = ctypes.py_object
dll._PyAsyncGenValueWrapperNew.argtypes = (ctypes.py_object,)

@types.coroutine
def yield_native(val):
    yield dll._PyAsyncGenValueWrapperNew(val)

def native_async_generator(afn):
    @functools.wraps(afn)
    async def wrapper(*a, **kw):
        if False:
            yield None   # force this to be an async generator
        await afn(*a, **kw)
    return wrapper

@native_async_generator
async def agen3():
    for i in range(N):
        await yield_native(i)

# another hypothetical way, more Windows-friendly
async def wrapper_maker():
    value = None
    while True:
        value = yield value
# get an async generator object out
wrapper_maker = wrapper_maker()
# transmute it to a regular generator object
type_p = ctypes.c_ulong.from_address(id(wrapper_maker) + ctypes.sizeof(ctypes.c_ulong))
assert type_p.value == id(types.AsyncGeneratorType)
type_p.value = id(types.GeneratorType)
# now wrapper_maker.send(x) returns an AsyncGenWrappedValue of x
wrapper_maker.send(None)

@types.coroutine
def yield_native_2(val):
    yield wrapper_maker.send(val)

@native_async_generator
async def agen4():
    for i in range(N):
        await yield_native_2(i)

# async iterator
class AIter:
    def __init__(self):
        self.i = 0

    def __aiter__(self):
        return self

    async def __anext__(self):
        i = self.i
        if i >= N:
            raise StopAsyncIteration
        self.i += 1
        return i
```
results:
```
In [2]: %time run(test(agen()))     # native asyncgen
CPU times: user 1.5 s, sys: 0 ns, total: 1.5 s
Wall time: 1.5 s
Out[2]: 49999995000000

In [3]: %time run(test(agen2()))    # @async_generator
CPU times: user 39.5 s, sys: 0 ns, total: 39.5 s
Wall time: 39.5 s
Out[3]: 49999995000000

In [4]: %time run(test(agen3()))    # @native_async_generator + yield_native
CPU times: user 8.63 s, sys: 3.33 ms, total: 8.63 s
Wall time: 8.63 s
Out[4]: 49999995000000

In [5]: %time run(test(AIter()))    # async iterator
CPU times: user 4.27 s, sys: 0 ns, total: 4.27 s
Wall time: 4.27 s
Out[5]: 49999995000000

In [6]: %time run(test(agen4()))    # @native_async_generator + yield_native_2
CPU times: user 4.83 s, sys: 0 ns, total: 4.83 s
Wall time: 4.83 s
Out[6]: 49999995000000
```
Would you be receptive to a patch that makes @async_generator work like agen4 in this example on CPython 3.6+? I know the "transmute an async generator to a regular generator" operation is sketchy, but I looked through genobject.c and I'm pretty well convinced it's safe -- an async generator object is a generator object with a few extra fields tacked on at the end, only one of which is an object pointer (ag_finalizer) and that's NULL until the first call to asend()/athrow()/etc (which we never invoke on our wrapper_maker).