@njsmith 2018-03-05
Wow, that's super clever. I like that it's much simpler than the current version of this hack, and that it can be motivated as an invisible optimization. (Maybe we wouldn't even want to document that `yield_from_` works in native async generators.)

I suspect you'll run into a bit of trouble getting that to pass the current test suite, because we have some tests to check that we don't suffer from [bpo-32526](https://bugs.python.org/issue32526), and native async generators *do* suffer from that bug. Someone should probably fix that at some point... But I guess merely working as well as native async generators would be OK, and we could mark that test as `xfail` for now (on configurations where the hack is used).

I think this will also cause behavioral differences with respect to the async generator GC hooks. Currently `@async_generator` functions never interact with those hooks (unlike native async generators), and I think with your hack it will start using the GC hooks. This is arguably the right thing. There's some discussion in #13. It's definitely complicated enough that we'd want to add tests for it though! And once PyPy gets native async generator support then we'll probably need to also implement GC hook support inside `AsyncGenerator`, since it won't be using this hack.